Move roadmatrix out of folder + rename to roadmatrix
