+++ CMakeLists.txt
@@ -9,7 +9,7 @@ set(SOURCES
     "src/street_environment.cpp"
     "src/dynamic_entity.cpp"
     "src/obstacle.cpp"
-    "src/road_matrix/road_matrix.cpp"
+    "src/roadmatrix.cpp"
 )
 
 set(HEADERS
@@ -22,7 +22,7 @@ set(HEADERS
     "include/street_environment/car.h"
     "include/street_environment/dynamic_entity.h"
     "include/street_environment/pose.h"
-    "include/street_environment/road_matrix/road_matrix.h"
+    "include/street_environment/roadmatrix.h"
 )
 
 #set compiler flags
@@ -52,7 +52,7 @@ target_link_libraries(street_environment PRIVATE lms_math lmscore)
 
 if(LOCAL_GMOCK)
 set(TESTS
-    "src/test/road_matrix_test.cpp"
+    "src/test/roadmatrix_test.cpp"
 )
 add_executable(street_environment_test ${TESTS})
 target_link_libraries(street_environment_test PRIVATE
